requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
DateTime==5.5
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
setuptools==76.0.0
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
zope.interface==7.2
